requirements.txt
slack-bolt
slack-sdk
PyGithub
pandas
streamlit
plotly